ogg ファイルの読み込みに失敗するケースが多々あった不具合を修正。（読み込みの高速化）
BASSマニュアルでは、BASS_ChannelGetLength() で返されるサイズは「OGGファイルの場合、返される長さは通常正確」と書いてあるが、実際には正確にはならない。
正確に同じサイズをデコードできなかった場合には、エラーを返していた。

+++ FDK/Code/03.Sound/Cxa.cs
@@ -68,7 +68,7 @@ namespace FDK
             return 0;
         }
 
-        public override int Decode( ref byte[] Dest, long offset )
+        public override long Decode( ref byte[] Dest, long offset )
         {
             #region [ Decodig xa data ]
             srcBuf = new byte[ format.Blocks * format.BlockSizeXa ];
