+++ FDK/Code/03.Sound/SoundDecoder.cs
@@ -9,7 +9,11 @@
         public CWin32.WAVEFORMATEX wfx { get; protected set; }
 
         public abstract int Open( string filename );
-        public abstract int Decode( ref byte[] Dest, long offset );
+        /// <returns>
+        /// <para lang="ja">実際にデコードされたバイト数。</para>
+        /// <para lang="en">The number of bytes actually decoded.</para>
+        /// </returns>
+        public abstract long Decode( ref byte[] Dest, long offset );
         public abstract void Close();
     }
 }
